+++ client/src/components/content/PropertiesPanel/PropertiesPanelView/PropertiesPanelComponents/PPApplicationDropdown.jsx
@@ -1,7 +1,7 @@
 /* eslint-disable react/prop-types */
 import React from 'react';
 import { Select } from 'antd';
-import styles from '../PropertiesPanel.module.css';
+import styles from '../../PropertiesPanel.module.css';
 
 const { Option } = Select;
 
@@ -10,7 +10,7 @@ const { Option } = Select;
  * @category Client
  * @component
  */
-const PropertiesPanelApplicationDropdown = ({
+const PPApplicationDropdown = ({
   onApplicationSelection,
   currentSelection,
   applications,
@@ -32,4 +32,4 @@ const PropertiesPanelApplicationDropdown = ({
   </>
 );
 
-export default PropertiesPanelApplicationDropdown;
+export default PPApplicationDropdown;
